refactor(ofm_core/contracts): extract termination cluster
Move the manager-initiated termination path and let-expire exit-intent flags out of
contracts/mod.rs into contracts/termination.rs: preview_contract_termination,
terminate_contract_now, set/clear_contract_exit_intent, has_let_expire_intent, plus
termination_severance_cost and contract_terminated_message. Byte-faithful move via
script; five public entry points re-exported via `pub use`. terminate_contract_now
still reaches ContractReleaseReason and release_player_contract (kept in mod.rs for
now) via `use super::*`.

387 ofm_core lib + 28 contracts integration tests pass; clippy clean in contracts/.
